requirements.txt
amqp==5.0.9
arrow==1.2.1
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
boto3==1.20.37
botocore==1.23.37
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
decouple==0.0.7
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.3
Django==4.0
django-allauth==0.47.0
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-formtools==2.3
django-jazzmin==2.4.8
django-phone-field==1.8.1
django-phone-verify==2.0.1
django-phonenumber-field==6.0.0
django-resized==0.3.11
django-ses==2.3.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
filelock==3.3.1
future==0.18.2
idna==3.3
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
kombu==5.2.3
Markdown==3.3.4
MarkupSafe==2.0.1
nexmo==2.5.2
oauthlib==3.1.1
packaging==21.3
pbr==5.6.0
phonenumbers==8.12.41
Pillow==8.4.0
platformdirs==2.4.0
prompt-toolkit==3.0.24
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
stevedore==3.5.0
twilio==7.4.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.8.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
